feat: 添加BTC split小数费率支持和auto-clock-in后台运行功能
- 修改BTC split命令支持小数费率（parseFloat替代parseInt）
- 新增完整的auto-clock-in后台运行管理系统
- 添加PM2配置文件和daemon管理脚本
- 新增systemd服务配置支持
- 更新package.json添加后台服务管理命令
- 添加详细的后台运行使用说明文档

+++ src/cli/btc.ts
@@ -106,7 +106,7 @@ export const btcSplit = new Command('split')
     console.log(
       await btc.splitUtxos({
         utxos: accountSpendableTotalUtxos,
-        feeRate: parseInt(options.feeRate),
+        feeRate: parseFloat(options.feeRate),
         account,
         provider,
         signer,
